+++ aws/services/init.go
@@ -30,7 +30,7 @@ var (
 	AccessService, InfraService, StorageService, MessagingService, DNSService, LambdaService, MonitoringService, CdnService, CloudformationService       cloud.Service
 	EksService, DynamodbService, SecretsmanagerService, ApigatewayService, SsmService, EfsService, CloudtrailService, CloudwatchlogsService              cloud.Service
 	ElasticacheService, EventbridgeService, StepfunctionsService, WafService, ConfigserviceService, KinesisService, RedshiftService, CodepipelineService cloud.Service
-	CodebuildService, BeanstalkService, CodedeployService, GlueService                                                                                   cloud.Service
+	CodebuildService, BeanstalkService, CodedeployService, GlueService, SesService                                                                       cloud.Service
 )
 
 func Init(profile, region string, extraConf map[string]any, log *logger.Logger, profileSetterCallback func(val string) error, enableNetworkMonitor bool) error {
@@ -75,6 +75,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	BeanstalkService = NewBeanstalk(cfg, profile, extraConf, log)
 	CodedeployService = NewCodedeploy(cfg, profile, extraConf, log)
 	GlueService = NewGlue(cfg, profile, extraConf, log)
+	SesService = NewSes(cfg, profile, extraConf, log)
 
 	cloud.ServiceRegistry[InfraService.Name()] = InfraService
 	cloud.ServiceRegistry[AccessService.Name()] = AccessService
@@ -105,6 +106,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	cloud.ServiceRegistry[BeanstalkService.Name()] = BeanstalkService
 	cloud.ServiceRegistry[CodedeployService.Name()] = CodedeployService
 	cloud.ServiceRegistry[GlueService.Name()] = GlueService
+	cloud.ServiceRegistry[SesService.Name()] = SesService
 
 	awsspec.CommandFactory = &awsspec.AWSFactory{
 		Log: log,